+++ tf_rnn/layers/utils/metrics.py
@@ -5,9 +5,6 @@
 
 from . import Accumulator
 
-# The following import is only needed for type hinting
-from ...logger import Logger
-
 
 class Metrics(object):
     """Holds the performance metrics for a given model.
@@ -18,16 +15,15 @@ class Metrics(object):
     - test (Accumulator): The accumulator that holds test performance metrics
     """
 
-    def __init__(self, logger: Logger, max_sequence_length: int):
+    def __init__(self, max_sequence_length: int):
         """Creates a metrics object with three empty accumulators.
 
         Params:
-        logger (logger.Logger): The logger from the model
-        max_sequence_length (int): The maximum sequence length for this dataset
+        - max_sequence_length (int): The maximum sequence length for this dataset
         """
-        self.train = Accumulator(logger, max_sequence_length)
-        self.valid = Accumulator(logger, max_sequence_length)
-        self.test = Accumulator(logger, max_sequence_length)
+        self.train = Accumulator(max_sequence_length)
+        self.valid = Accumulator(max_sequence_length)
+        self.test = Accumulator(max_sequence_length)
     # End of __init__()
 
     def advance(self):